// juniormagnus/types/quiz.ts
export interface ApiResponse<T> {
  success: boolean;
  message?: string;
  data?: T;
  count?: number;
  statusCode?: number;
}